export-service/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
httpx>=0.28.0
redis>=5.2.0
python-docx>=1.1.0
reportlab>=4.2.0
openpyxl>=3.1.0
jinja2>=3.1.0
